Don't store library state with globals
Don't store any of the state as global variables or limit the library
to just STDIN/STDOUT/STDERR. This allows for better handling of the
internal state of the library for integration into applications.

Try and fix the cursor position and some issues with the internal
buffering of the command line text. Some stuff is a bit messed up
still but at least most of the library acts like a terminal you would
expect.

Remove atexit() functionality as that should be handled by the
implementing application and not by the library itself. It's not our
job to make sure that LinenoiseFreeState and LinenoiseRestore is called.